backend/requirements.txt
# --- Core Backend ---
fastapi
uvicorn
python-multipart
python-dotenv

# --- AI & Embeddings ---
groq
google-genai
numpy
sentence-transformers
torch

# --- Vector Database ---
qdrant-client

# --- Parsing & Chunking ---
llama-index-core
unstructured[pdf]
pypdf

# --- Search & Ranking ---
rank-bm25